refactor/folder structure (#9)
* refactor:components-folder-changed-location

* refactor:favicon.ico-file-moved-to-public-folder

* refactor:vercel-files-deleted

* refactor:global-styles-file-moved-to-own-styles-folded
