+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "./globals.css";
-import NavBar from "./components/navbar/NavBar";
-import BottomTrim from "./components/navbar/NavBarBottomTrim";
+import "../styles/globals.css";
+import NavBar from "../components/navbar/NavBar";
+import BottomTrim from "../components/navbar/NavBarBottomTrim";
 import color from "@/lib/color";
 
 const inter = Inter({ subsets: ["latin"] });
