+++ components/navbar/NavBarBottomTrim.tsx
